Get agent_name from config file
1. VRouter Name
---------------
In current design, name of agent (vrouter-agent) registered with controller is
assumed to be same as hostname.

In case of OVS, there can potentially be multiple instances of tor-agent
running on same server. The name of agent will need to be different for
each instance. Making agent-name configurable.

2. VMI to VM link
------------------
VM-Interface is set to be INACTIVE, if VM for the interface is not known.
In case of Tor-agent, there is no VM object for VM-Interface.

Making VMI state independnet of VM for the interface